Integrated code into scaffolding. All data now sent via POST.
I integrated the tictactoe code into the yesod scaffolding and updated
the config files accordingly. I deleted the files from the example project
that are no longer needed.

The coordinates of the user's click in the field and the current position
are now transmitted exclusively via POST, except for the SVG generation. The
latter still gets the position via the URL.