Implement smarter AI.
Roughly speaking, the AI targets cells where there are more possibilities
for a ship to fit in. In addition, it follows a checkerboard pattern to
reduce the number of shots needed to find all ships.
This way, it needs less then 50 shots on average to sink all ships
(assuming standard rules). So over 60% are hits.
However, this strategy needs to be adapted for movable ships.