Finally fixed that bug... (in game control)
When determining which ship was fired at, don't consider the sunk ones because they are "gone".
The problem was this: Ship A is sunk. Ship B moves on top of ship A. If you fire at A, you
sometimes get the result for B (namely "Sunk") instead of the result for A (namely "Hit").
This confused the AI very much.